+++ module-api/src/main/kotlin/finn/service/TickerQueryService.kt
@@ -5,6 +5,7 @@ import finn.filter.TickerSearchFilter
 import finn.queryDto.TickerQueryDto
 import finn.repository.TickerRepository
 import org.springframework.stereotype.Service
+import java.math.BigDecimal
 import java.util.*
 
 @Service
@@ -26,7 +27,7 @@ class TickerQueryService(
         return tickerRepository.getTickerByTickerCode(tickerCode)
     }
 
-    suspend fun getYesterdayAtr(tickerId: UUID): Double {
+    suspend fun getYesterdayAtr(tickerId: UUID): BigDecimal {
         return tickerRepository.getPreviousAtrByTickerId(tickerId)
     }
 }
